+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 ##### Unreleased
+- [`Array` grouping proposal](https://github.com/tc39/proposal-array-grouping):
+  - Methods:
+    - `Object.groupBy` method
+    - `Map.groupBy` method
+  - Moved to stable ES, November 2023 TC39 meeting
+  - Added `es.` namespace modules, `/es/` and `/stable/` namespaces entries
 - Some minor `atob` / `btoa` improvements
 - Compat data improvements:
   - [`Promise.withResolvers`](https://github.com/tc39/proposal-promise-with-resolvers) marked as shipped from FF121